not pretty, but super basic outline is there

+++ app/src/main/java/com/example/myweatherapp/models/Wind.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 
 data class Wind(
     val deg: Int,
-    val gust: Int,
+    val gust: Double,
     val speed: Double
 )
